requirements.txt
pillow
fastapi
uvicorn
requests
pdfplumber
jinja2